requirements/development.txt
-r common.txt

# development only dependencies in alphabetical order
django-debug-toolbar
pre-commit
pre-commit-update
ruff
rich

pytest~=7.4.0
pytest-cov~=4.1.0
pytest-django~=4.5.2
pytest-dotenv~=0.5.2
pytest-mock~=3.11.1
